Moved shared helpers to utils.
